Refactor folder layout and module exports for main addon
