+++ test-app/tests/integration/components/headless-form-data-test.gts
@@ -13,7 +13,7 @@ import {
 } from '@ember/test-helpers';
 import { module, skip, test } from 'qunit';
 
-import HeadlessForm from 'ember-headless-form/components/headless-form';
+import { HeadlessForm } from 'ember-headless-form';
 import sinon from 'sinon';
 import { setupRenderingTest } from 'test-app/tests/helpers';
