+++ test-app/tests/integration/components/headless-form-glint-test.gts
@@ -4,7 +4,7 @@
 import { render } from '@ember/test-helpers';
 import { module, test } from 'qunit';
 
-import HeadlessForm from 'ember-headless-form/components/headless-form';
+import { HeadlessForm } from 'ember-headless-form';
 import { setupRenderingTest } from 'test-app/tests/helpers';
 
 module('Integration Component HeadlessForm > Glint', function (hooks) {
