+++ test-app/tests/integration/components/headless-form-validation-test.gts
@@ -5,7 +5,7 @@ import { tracked } from '@glimmer/tracking';
 import { blur, click, fillIn, render, rerender } from '@ember/test-helpers';
 import { module, test } from 'qunit';
 
-import HeadlessForm from 'ember-headless-form/components/headless-form';
+import { HeadlessForm } from 'ember-headless-form';
 import sinon from 'sinon';
 import { setupRenderingTest } from 'test-app/tests/helpers';
 
@@ -15,7 +15,7 @@ import type {
   FieldValidateCallback,
   ErrorRecord,
   ValidationError,
-} from 'ember-headless-form/components/-private/types';
+} from 'ember-headless-form';
 
 import { input } from '../../helpers/dom';
